Sat, Jul 29, 2023, 9:27 PM -03:00

+++ scrabble/scrabble.c
@@ -38,7 +38,7 @@ int compute_score(string word)
     {
         for(int j = 0; j < len; j++)
         {
-            if((int)word[i] == alp[j])
+            if(word[i] == alp[j])
             {
                 score += POINTS[j];
             }
